requirements.txt
fastapi
uvicorn[standard]
gradio
sentence-transformers
PyPDF2
numpy
PyYAML
requests
tqdm
faiss-cpu
llama-cpp-python